feat: add useCheckLoginState useHooks on Main page
